Moved `useManager` hook
