+++ app/components/GettingStartedModal/common/managerHooks.ts
@@ -1,9 +1,9 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import _ from 'lodash';
 
 import Internal from '../../../utils/Internal';
 import Manager from '../../../utils/Manager';
+import useManager from '../../../utils/hooks/useManager';
 
 import type { ReduxState } from '../../../reducers';
 
@@ -11,15 +11,6 @@ const getCurrentDistribution = (manager: Manager) => {
     const currentDistributionName = manager.getDistributionName().trim();
     const currentDistributionRelease = manager.getDistributionRelease().trim();
     return `${currentDistributionName.toLowerCase()} ${currentDistributionRelease.toLowerCase()}`;
-};
-
-/**
- * Gets current manager from context.
- * @returns current manager object
- */
-export const useManager = () => {
-    const manager = useSelector((state: ReduxState) => state.manager);
-    return useMemo(() => new Manager(manager), [manager]);
 };
 
 /**
