+++ app/components/GettingStartedModal/steps/SummaryStep/SummaryStep.tsx
@@ -10,7 +10,8 @@ import {
     useSecretsInstallationTasks,
     useBlueprintsInstallationTasks
 } from '../../installation/tasks';
-import { useInternal, useManager } from '../../common/managerHooks';
+import { useInternal } from '../../common/managerHooks';
+import useManager from '../../../../utils/hooks/useManager';
 import { createResourcesInstaller, TaskDetails, TaskStatus, TaskType } from '../../installation/process';
 import PluginsInstallationTasks from './PluginsInstallationTasks';
 import SecretsInstallationTasks from './SecretsInstallationTasks';
